backend/requirements.txt
asgiref==3.8.1
Django==4.2.18
django-cors-headers==4.6.0
django-filter==24.3
django-multiselectfield==0.1.13
djangorestframework==3.15.2
importlib_metadata==8.6.1
Markdown==3.7
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
zipp==3.21.0
pdfplumber==0.11.6
python-docx==1.1.2
googletrans==4.0.2